pyproject.toml
[project]
name = "djangoprojecttest"
version = "0.1.0"
description = ""
authors = [
    {name = "bratxemss",email = "bratxems@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "celery (>=5.5.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
